New ClassDAO

+++ src/main/java/daos/ClassDAO.java
@@ -7,6 +7,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.UUID;
 import utils.DBUtils;
 
 public class ClassDAO {
@@ -17,7 +18,6 @@ public class ClassDAO {
 
     private static void closeConnection() {
         try {
-
             if (rs != null) {
                 rs.close();
             }
@@ -42,14 +42,15 @@ public class ClassDAO {
 
         try {
             conn = DBUtils.makeConnection();
+
             String sql = "SELECT u.userId, u.fullName, u.email, u.password, u.role, cl.classId, cl.name, cl.teacherId  "
                     + "FROM (User u right join User_Class ucl on u.userId = ucl.userId) left join Class cl on ucl.classId = cl.classId "
                     + "WHERE cl.teacherId = ? AND u.role = 'student' "
-                    + "ORDER BY cl.classId";
+                    + "ORDER BY cl.classId ";
             preStm = conn.prepareStatement(sql);
             preStm.setString(1, teacherId);
             rs = preStm.executeQuery();
-            
+
             if (rs.next()) {
                 //Tạo biến chứa giá trị classId của row hiện tại
                 String classId = rs.getString("cl.classId");
@@ -60,25 +61,25 @@ public class ClassDAO {
                     String email = rs.getString("u.email");
                     String password = rs.getString(" u.password");
                     String role = rs.getString("u.role");
-                    
+
                     //Phân loại lớp
                     String newClassId = rs.getString("cl.classId");
                     //So sánh row đang xét có trùng classId với row trên không
-                    if(newClassId != classId){
+                    if (newClassId.equals(classId)) {
+                        listUser.add(new User(userId, fullName, email, password, role));
+                    } else {
                         classId = newClassId;
                         list.add(new Class(rs.getString("cl.classID"), teacherId, rs.getString("cl.name"), listUser));
-                        
+
                         //Add vô Student đầu tiên vô class mới
                         listUser = null;
-                        listUser.add(new User(userId, fullName, email, password, role));
-                    }else{
                         listUser.add(new User(userId, fullName, email, password, role));
                     }
                 }
             }
-            return list;
         } catch (Exception e) {
             e.printStackTrace();
+            return null;
         } finally {
             closeConnection();
         }
@@ -95,14 +96,14 @@ public class ClassDAO {
 
         try {
             conn = DBUtils.makeConnection();
-            String sql = "SELECT cl.classId, cl.teacheId, cl.name , u.userId, u.fullName, u.email, u.password, u.role"
+            String sql = "SELECT cl.classId, cl.teacherId, cl.name , u.userId, u.fullName, u.email, u.password, u.role"
                     + "FROM ( Class cl left join User_Class ucl on cl.classId = ucl.classId ) inner join User u on ucl.userId = u.userId  "
                     + "WHERE ucl.userId = ? AND u.role = 'student'  "
                     + "ORDER BY cl.classId";
             preStm = conn.prepareStatement(sql);
             preStm.setString(1, studentId);
             rs = preStm.executeQuery();
-    
+
             if (rs.next()) {
                 //Tạo biến chứa giá trị classId của row hiện tại
                 String classId = rs.getString("cl.classId");
@@ -113,26 +114,25 @@ public class ClassDAO {
                     String email = rs.getString("u.email");
                     String password = rs.getString(" u.password");
                     String role = rs.getString("u.role");
-                  
-                    
-                     //Phân loại lớp
+
+                    //Phân loại lớp
                     String newClassId = rs.getString("cl.classId");
                     //So sánh row đang xét có trùng classId với row trên không
-                    if(newClassId != classId){
+                    if (newClassId.equals(classId)) {
+                        listUser.add(new User(userId, fullName, email, password, role));
+                    } else {
                         classId = newClassId;
                         list.add(new Class(rs.getString("cl.classID"), rs.getString("cl.teacheId"), rs.getString("cl.name"), listUser));
-                        
+
                         //Add vô Student đầu tiên vô class mới
                         listUser = null;
                         listUser.add(new User(userId, fullName, email, password, role));
-                    }else{
-                        listUser.add(new User(userId, fullName, email, password, role));
-                    }  
+                    }
                 }
             }
-            return list;
         } catch (Exception e) {
             e.printStackTrace();
+            return null;
         } finally {
             closeConnection();
         }
@@ -140,42 +140,52 @@ public class ClassDAO {
     }
 
     // Create a new class that taught by teacher A
-    public static Class createNewClass(Class c, String teacherId) throws SQLException {
+    public static Class createNewClass(Class c, String teacherId) {
         conn = null;
         preStm = null;
         rs = null;
-        Class classResult = null;
-        ArrayList<User> listUser = new ArrayList<>();
+
         try {
             //Create a Class
             conn = DBUtils.makeConnection();
-            String sql = "INSERT INTO Class "
-                    + "VAVLUES( ?, ?, ?)";
-            preStm = conn.prepareStatement(sql);
-            preStm.setString(1, c.getClassId());
-            preStm.setString(2, teacherId);
-            preStm.setString(3, c.getName());
-            rs = preStm.executeQuery();
 
-            //Insert The Student into class
-            preStm = null;
-            rs = null;
-            for (User user : c.getStudents()) {
-                sql = "INSERT INTO USER_CLASS "
-                        + "VALUES(? , ?)";
+            if (conn != null) {
+                conn.setAutoCommit(false);
+                c.setTeacherId(teacherId);
+                c.setClassId(UUID.randomUUID().toString());
+
+                //Insert Class
+                String sql = "INSERT INTO Class(classId, teacherId, name) "
+                        + "VALUES( ?, ?, ?)";
+
                 preStm = conn.prepareStatement(sql);
-                preStm.setString(1, user.getUserId());
-                preStm.setString(2, c.getClassId());
-                listUser.add(new User(user.getUserId(), user.getFullName(), user.getEmail(), user.getPassword(), user.getRole()));
+                preStm.setString(1, c.getClassId());
+                preStm.setString(2, c.getTeacherId());
+                preStm.setString(3, c.getName());
+                rs = preStm.executeQuery();
+
+                //Insert The Student into class
+                preStm = null;
+                for (User user : c.getStudents()) {
+                    try {
+                        sql = "INSERT INTO USER_CLASS(userId, classId) "
+                                + "VALUES(? , ?)";
+                        preStm = conn.prepareStatement(sql);
+                        preStm.setString(1, user.getUserId());
+                        preStm.setString(2, c.getClassId());
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                }
             }
-            classResult = new Class(c.getClassId(), teacherId, c.getName(), listUser);
-            return classResult;
+            conn.commit();
+            conn.setAutoCommit(true);
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
             closeConnection();
         }
-        return classResult;
+        return c;
     }
 
     // A student attends a new class
@@ -186,15 +196,17 @@ public class ClassDAO {
         int result = 0;// Nếu Hàm trả về 0 thì sẽ không có dòng nào ảnh hưởng
         try {
             conn = DBUtils.makeConnection();
-            String sql = "INSERT INTO User_Class "
-                    + "Values( ?, ?)";
-            preStm = conn.prepareStatement(sql);
-            preStm.setString(1, studentId);
-            preStm.setString(2, c.getClassId());
 
-            rs = preStm.executeQuery();
-            result = preStm.executeUpdate();
-            return result;
+            if (conn != null) {
+                String sql = "INSERT INTO User_Class(userId, classId)"
+                        + "Values( ?, ?)";
+                preStm = conn.prepareStatement(sql);
+                preStm.setString(1, studentId);
+                preStm.setString(2, c.getClassId());
+
+                rs = preStm.executeQuery();
+                result = preStm.executeUpdate();
+            }
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
@@ -211,38 +223,46 @@ public class ClassDAO {
         String sql = null;
         int result = 0;// Nếu Hàm trả về 0 thì sẽ không có dòng nào ảnh hưởng
         try {
+            conn = DBUtils.makeConnection();
 
-            //Update the NAME of Class
-            sql = "UPDATE Class "
-                    + "SET name = ? "
-                    + "WHERE classId = ?";
-            preStm = conn.prepareStatement(sql);
-            preStm.setString(1, c.getName());
-            preStm.setString(2, c.getClassId());
-            rs = preStm.executeQuery();
-
-            //Delete old menbers in a Class
-            preStm = null;
-            rs = null;
-            sql = "DELETE FROM User_Class "
-                    + "Where classId = ? ";
-            preStm = conn.prepareStatement(sql);
-            preStm.setString(1, c.getClassId());
-            rs = preStm.executeQuery();
-
-            //INSERT NEW data in CLASS
-            preStm = null;
-            rs = null;
-            for (User user : c.getStudents()) {
-                sql = "INSERT INTO User_Class "
-                        + "Values( ?, ?) ";
+            if (conn != null) {
+                conn.setAutoCommit(false);
+                //Update the Class
+                sql = "UPDATE Class "
+                        + "SET name = ? "
+                        + "WHERE classId = ? ";
                 preStm = conn.prepareStatement(sql);
-                preStm.setString(1, user.getUserId());
+                preStm.setString(1, c.getName());
                 preStm.setString(2, c.getClassId());
-                rs = preStm.executeQuery();
                 result = preStm.executeUpdate();
+
+                //Delete old menbers in a Class
+                preStm = null;
+                rs = null;
+                sql = "DELETE FROM User_Class "
+                        + "Where classId = ? ";
+                preStm = conn.prepareStatement(sql);
+                preStm.setString(1, c.getClassId());
+                preStm.executeUpdate();
+
+                //INSERT NEW data in User_Class
+                preStm = null;
+                rs = null;
+                for (User user : c.getStudents()) {
+                    try {
+                        sql = "INSERT INTO User_Class(userId, classId) "
+                                + "Values( ?, ?) ";
+                        preStm = conn.prepareStatement(sql);
+                        preStm.setString(1, user.getUserId());
+                        preStm.setString(2, c.getClassId());
+                        preStm.executeUpdate();
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                }
             }
-            return result;
+            conn.commit();
+            conn.setAutoCommit(true);
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
@@ -260,23 +280,36 @@ public class ClassDAO {
         String sql = null;
         try {
             conn = DBUtils.makeConnection();
-            //Delete All Student in Clas
-            sql = "DELETE FROM User_Class "
-                    + "WHERE classId = ? ";
-            preStm = conn.prepareStatement(sql);
-            preStm.setString(1, classId);
-            rs = preStm.executeQuery();
 
-            //Delete Class
-            preStm = null;
-            sql = "DELETE FROM Class "
-                    + "WHERE classId = ? ";
-            preStm.setString(1, classId);
-            rs = preStm.executeQuery();
-            result = preStm.executeUpdate();
-            return result;
+            if (conn != null) {
+                conn.setAutoCommit(false);
+                //Delete All Student in Class
+                sql = "DELETE FROM User_Class "
+                        + "WHERE classId = ? ";
+                preStm = conn.prepareStatement(sql);
+                preStm.setString(1, classId);
+                preStm.executeQuery();
+
+                //Delete Class
+                preStm = null;
+                sql = "DELETE FROM Class "
+                        + "WHERE classId = ? ";
+                preStm = conn.prepareStatement(sql);
+                preStm.setString(1, classId);
+                result = preStm.executeUpdate();
+                return result;
+            }
+            conn.commit();
+            conn.setAutoCommit(true);
         } catch (Exception e) {
+            try {
+                conn.rollback();
+            } catch (Exception e1) {
+                e1.printStackTrace();
+                return result;
+            }
             e.printStackTrace();
+            return result;
         } finally {
             closeConnection();
         }
@@ -299,9 +332,9 @@ public class ClassDAO {
 
             rs = preStm.executeQuery();
             result = preStm.executeUpdate();
-            return result;
         } catch (Exception e) {
             e.printStackTrace();
+            return result;
         } finally {
             closeConnection();
         }
